+++ src/pages/*.ts
@@ -1,6 +1,6 @@
 import Elysia from "elysia";
 import { home } from "./(home)";
-import { admin } from "./admin";
+import { Admin } from "./admin/admin";
 import { leaderboard } from "./leaderboard";
 import { match } from "./log-match/logMatch";
 import { login } from "./login";
@@ -15,7 +15,7 @@ const publicPages = new Elysia()
   .use(leaderboard)
   .use(stats)
   .use(profile)
-  .use(admin)
+  .use(Admin)
   .use(matchResult);
 
 const authPages = new Elysia().use(play).use(match);
